requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
Pillow>=10.0
opencv-python>=4.8
torch>=2.2
torchvision>=0.17
adjustText>=1.1
geopandas>=0.14
rasterio>=1.3
shapely>=2.0
openpyxl>=3.1
